+++ UnitTests/CombinedUnits/RatioTests.VolumeConcentration.cs
@@ -64,24 +64,4 @@ public partial class RatioTests
         Assert.AreEqual(20, WorkingCompares);
 
     }
-
-    [TestMethod]
-    public void FromVolume01()
-    {
-        double? value = null;
-
-        var v = EngineeringUnits.Volume.From(value, null);
-        Assert.IsNull(v);
-
-    }
-    [TestMethod]
-    public void FromVolume02()
-    {
-        double? value = 100;
-        var v = EngineeringUnits.Volume.From(value, VolumeUnit.CubicFoot);
-        var v2 = EngineeringUnits.Volume.From(value, null);
-        Assert.IsNotNull(v);
-        Assert.IsNull(v2);
-
-    }
 }
